refactor: merge portfolio review job into workflow
