+++ backend/test/agentCreateAsync.test.ts
@@ -5,7 +5,7 @@ import { setAiKey, setBinanceKey } from '../src/repos/api-keys.js';
 const reviewAgentPortfolioMock = vi.fn<
   (log: unknown, agentId: string) => Promise<unknown>
 >(() => new Promise(() => {}));
-vi.mock('../src/jobs/review-portfolio.js', () => ({
+vi.mock('../src/workflows/portfolio-review.js', () => ({
   reviewAgentPortfolio: reviewAgentPortfolioMock,
 }));
